Time: 1901 ms (23.59%), Space: 447.2 MB (5.13%) - LeetHub

+++ 3250-maximum-square-area-by-removing-fences-from-a-field/3250-maximum-square-area-by-removing-fences-from-a-field.cpp
@@ -7,8 +7,8 @@ public:
 
         unordered_set<int>sides;
 
-        for(int i=0; i<fencePositions.size(); i++){
-            for(int j=i+1; j<fencePositions.size(); j++){
+        for(size_t i=0; i<fencePositions.size(); i++){
+            for(size_t j=i+1; j<fencePositions.size(); j++){
                 int side = fencePositions[j] - fencePositions[i];
                 sides.insert(side);
             }
